+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Things under "main" are not released yet.
 
+## main
+
+- Error on `==` where both sides are functions.
+
 ## v0.9.4
 
 - Set the current dir of `bin/jsonnetfmt` to the workspace root.